:recycle: refactor(#168/global): global 설정 클래스 ninedot-global 모듈로 이동
