+++ ninedot-global/src/main/java/org/sopt36/ninedotserver/config/auth/AuthConfig.java
@@ -1,4 +1,4 @@
-package org.sopt36.ninedotserver.global.config.auth;
+package org.sopt36.ninedotserver.config.auth;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
@@ -14,7 +14,7 @@ public class AuthConfig {
         factory.setConnectTimeout(5000); // 연결 타임아웃 5초 (5000ms) - 서버랑 연결하는데 5초 준다.
         factory.setReadTimeout(10000);   // 읽기 타임아웃 10초 (10000ms) - 데이터 불러오는데 10초 준다.
         return RestClient.builder()
-                   .requestFactory(factory)
-                   .build();
+            .requestFactory(factory)
+            .build();
     }
 }
