+++ ninedot-global/src/main/java/org/sopt36/ninedotserver/config/gemini/GeminiConfig.java
@@ -1,4 +1,4 @@
-package org.sopt36.ninedotserver.global.config.gemini;
+package org.sopt36.ninedotserver.config.gemini;
 
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
